+++ src/main/kotlin/icu/windea/pls/model/codeInsight/ParadoxImageCodeInsightContext.kt
@@ -141,8 +141,7 @@ data class ParadoxImageCodeInsightContext(
         
         private fun isMissing(iconPath: String, project: Project, context: PsiElement): Boolean {
             val iconSelector = fileSelector(project, context)
-            val iconFile = ParadoxFilePathSearch.searchIcon(iconPath, iconSelector).findFirst()
-            val missing = iconFile == null
+            val missing = ParadoxFilePathSearch.searchIcon(iconPath, iconSelector).findFirst() == null
             return missing
         }
     }