+++ src/main/kotlin/icu/windea/pls/script/codeInsight/completion/ParadoxScriptedVariableCompletionProvider.kt
@@ -38,7 +38,6 @@ class ParadoxScriptedVariableCompletionProvider : CompletionProvider<CompletionP
     private fun processScriptedVariable(context: ProcessingContext, result: CompletionResultSet, element: ParadoxScriptScriptedVariable): Boolean {
         ProgressManager.checkCanceled()
         val name = element.name ?: return true
-        if (context.completionIds?.add(name) == false) return true //排除重复项
         val tailText = element.value?.let { " = $it" }
         val typeFile = element.containingFile
         val lookupElement = LookupElementBuilder.create(element, name)