+++ CHANGELOG.md
@@ -205,7 +205,7 @@
 * [X] Optimization: Determine override order for targets (definitions, localisations, etc.) based on CWT configs files (Can be customized, see reference documentation for details)
 * [X] Optimization: For localisation file, it's unnecessary to escape double quotes in localisation text
 * [X] Optimization: It a target cannot be resolved, but related extended CWT configs exist, related code inspection can be configured to be ignored
-* [X] Optimization: If the scope context can be inferred from extended CWT configs, do not continue to be inferred from usages  
+* [X] Optimization: If the scope context can be inferred from extended CWT configs, do not continue to be inferred from usages
 * [X] Other optimizations and bug fixes
 
 ## 1.3.6
@@ -738,8 +738,7 @@ mult = modifier:$MODIFIER$ # 脚本参数作为某个复杂表达式中的整个
 
 * 更新到IDEA版本2023.1
 * 优化索引时的规则匹配逻辑
-*
-修复：[只能读取动态肖像而不能读取静态肖像 #13](https://github.com/DragonKnightOfBreeze/Paradox-Language-Support/issues/13)
+* 修复：[只能读取动态肖像而不能读取静态肖像 #13](https://github.com/DragonKnightOfBreeze/Paradox-Language-Support/issues/13)
 
 ## 0.9.3